refactor: a huge refactor
