+++ test/routingHeader.ts
@@ -1,4 +1,5 @@
-/* Copyright 2017, Google Inc.
+/*
+ * Copyright 2019 Google LLC
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
