+++ tsconfig.json
@@ -8,6 +8,7 @@
   },
   "include": [
     "src/*.ts",
+    "src/*/*.ts",
     "test/*.ts",
     "system-test/*.ts"
   ]
